Refactoring, change task colors, fix issues when navigating between vsmObjects and looking at a specific task etc...
