+++ interfaces/taskObject.tsx
@@ -11,6 +11,7 @@ export interface taskObject {
   solved?: boolean;
   solvedDate?: string;
   fkProject: number;
+  dsiplayIndex: string; // Todo: fix after this is resolved https://equinor-sds-si.atlassian.net/browse/VSM-67?focusedCommentId=25963
   objects?: [
     {
       fkObject: vsmObject["vsmObjectID"];
