+++ pages/projects/[id].tsx
@@ -50,7 +50,7 @@ export default function Project() {
         </Head>
 
         <main className={commonStyles.main}>
-          <Typography variant="h1">Error loading project {id}</Typography>
+          <Typography variant="h1">{error.message}</Typography>
           <p>
             We have some troubles with this VSM. Please try to refresh the page.
           </p>
