+++ utils/taskSorter.ts
@@ -1,0 +1,6 @@
+import { taskObject } from "../interfaces/taskObject";
+
+export const taskSorter = () => (a: taskObject, b: taskObject): number => {
+  if (!!a && !!b) return a.fkTaskType - b.fkTaskType;
+  return 1;
+};
